version 0.3.1

+++ connect/DeleteLocalMessages.go
@@ -8,8 +8,12 @@ import (
 func DeleteLocalMessages(slice []string) {
     settings.Mutex.Lock()
     for _, user := range slice {
-        _, err := settings.DataBase.Exec("DELETE FROM Local" + user)
-        utils.CheckError(err)
+        for _, check := range settings.User.Connections {
+            if check == user { 
+                _, err := settings.DataBase.Exec("DELETE FROM Local" + user)
+                utils.CheckError(err)
+            }
+        }
     }
     settings.Mutex.Unlock()
 }
